# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Miracle Birds API
  description: |
    # Miracle Birds - AI Intelligence Layer for CRM
    
    The Secure AI Intelligence Layer for Every CRM. Miracle Birds provides AI-powered customer intelligence, predictive analytics, and workflow automation on top of existing CRM platforms.
    
    ## Features
    - 🤖 AI CRM Copilot
    - 📊 Customer 360 Intelligence
    - 📈 Predictive Analytics (Churn, Lead Scoring, Revenue)
    - 🔗 CRM Integration (Salesforce, Zoho, HubSpot, Dynamics, Pipedrive)
    - 🔒 Enterprise Security & Compliance
    - 🚀 Workflow Automation
    
    ## Authentication
    All API endpoints (except auth endpoints) require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - **Standard Plan**: 1,000 requests/hour
    - **Professional Plan**: 5,000 requests/hour
    - **Enterprise Plan**: Unlimited
    
    ## Base URLs
    - **Production**: `https://api.miraclebirds.ai`
    - **Staging**: `https://api-staging.miraclebirds.ai`
    - **Development**: `http://localhost:8000`
    
  version: 1.0.0
  contact:
    name: Miracle Birds API Support
    email: api@miraclebirds.ai
    url: https://miraclebirds.ai/support
  license:
    name: Proprietary
    url: https://miraclebirds.ai/terms
servers:
  - url: https://api.miraclebirds.ai
    description: Production server
  - url: https://api-staging.miraclebirds.ai
    description: Staging server
...
paths:
...
  /api/v1/customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: Get customer
      description: Get customer details by ID
      operationId: getCustomer
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Customer ID
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Customers
      summary: Update customer
      description: Update customer information
      operationId: updateCustomer
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Customers
      summary: Delete customer
      description: Soft delete a customer
      operationId: deleteCustomer
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        detail:
          type: string
          example: Email is required
        code:
          type: string
          example: validation_error
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        external_id:
          type: string
        crm_source:
          type: string
          enum: [salesforce, zoho, hubspot, dynamics, pipedrive]
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        company:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [active, inactive, churned]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CustomerUpdate:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        company:
          type: string
        title:
          type: string
        status:
          type: string
    CustomerDetail:
      allOf:
        - $ref: '#/components/schemas/Customer'
        - type: object
          properties:
            health_score:
              type: number
              format: float
              example: 85.5
            churn_risk:
              type: string
              enum: [low, medium, high]
            lead_score:
              type: integer
              example: 92
            lifetime_value:
              type: number
              format: float
              example: 25000.00
            interactions_count:
              type: integer
              example: 45
...